Added note about type:service + spotcast
I find it pretty difficult to spot the type:service option, and get it working with my Google Setup. I sadly wish to have Sonos around - however, with 6 Google devices, i'm not going to swap them, anytime soon.

So for everyon with Google, i guess this could be informative for other people, too.

Thanks for a great solution, btw :-) I guess you're taking my 1. place for the best custom module - that's for sure!