tts - google_translate platform issue
The else path in line `this.hass.callService('tts', `${config.platform}_say`, opts);` is resulting in the service tts\google_translate_say being called if the platform is google_translate which does not work. I believe the correct service to call for tts is tts\google_say and it works when I change the code locally. Code has been modified to reflect to add a new `else if` path if the platform is google_translate.